Rename files missing '.component' (#1023)
