+++ packages/datagateway-search/src/searchPageCardView.component.tsx
